+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+* Additional error handling
+* New tests
+
+### Changes
+* New max dimensions, to stop heap out of memory errors
 
 ## [2.0.0] - 2020-04-21
 ### Added